==> src/12 gradient_boosting_machines/07_model_interpretation.py <==
# Long-Short Strategy, Part 4: How to interpret GBM results
# Understanding why a model predicts a certain outcome is very important for several reasons, including trust,
# actionability, accountability, and debugging.
# Insights into the nonlinear relationship between features and the outcome uncovered by the model, as well as
# interactions among features, are also of value when the goal is to learn more about the underlying drivers of
# the phenomenon under study.

from pathlib import Path
import warnings
from random import randint
import joblib
from itertools import product

import numpy as np
import pandas as pd

import shap
import lightgbm as lgb
from sklearn.inspection import plot_partial_dependence, partial_dependence

import matplotlib.pyplot as plt
from matplotlib.ticker import FuncFormatter
from mpl_toolkits.mplot3d import Axes3D
import seaborn as sns

np.random.seed(42)
idx = pd.IndexSlice
sns.set_style("whitegrid")
plt.rcParams["figure.dpi"] = 300
plt.rcParams["font.size"] = 14
warnings.filterwarnings("ignore")
pd.options.display.float_format = "{:,.2f}".format

YEAR = 252
DATA_STORE = Path("../data/assets.h5")


if __name__ == "__main__":
    ## Retrieve best model parameters
    # The results used below are created by the notebooks [gbm_trading_signals](05_gbm_trading_signals.ipynb).
    with pd.HDFStore("../data/12_data.h5") as store:
        best_params = store["best_params"]
    print(best_params)

    ## Get Training Data
    data = pd.read_hdf("../data/12_data.h5", "model_data").sort_index()
    data = data.loc[idx[:, "2013":"2018"], :]
    data.info()

    ## Create Binary Data
    dates = sorted(data.index.get_level_values("date").unique())
    train_dates = dates[-int(best_params.train_length) :]
    data = data.loc[idx[:, train_dates], :]

    labels = sorted(data.filter(like="_fwd").columns)
    features = data.columns.difference(labels).tolist()

    lookahead = 1
    label = f"r{lookahead:02}_fwd"
    categoricals = ["year", "month", "sector", "weekday"]

    lgb_train = lgb.Dataset(
        data=data[features],
        label=data[label],
        categorical_feature=categoricals,
        free_raw_data=False,
    )

    ## Train LightGBM Model
    params = dict(boosting="gbdt", objective="regression", verbose=-1)
    train_params = ["learning_rate", "num_leaves", "feature_fraction", "min_data_in_leaf"]
    params.update(best_params.loc[train_params].to_dict())
    for p in ["min_data_in_leaf", "num_leaves"]:
        params[p] = int(params[p])
    print(params)

    lgb_model = lgb.train(
        params=params, train_set=lgb_train, num_boost_round=int(best_params.boost_rounds)
    )

    ## Compute Feature Importance
    # A common approach to gaining insights into the predictions made by tree ensemble methods, such as gradient
    # boosting or random forest models, is to attribute feature importance values to each input variable. These feature
    # importance values can be computed on an individual basis for a single prediction or globally for an entire dataset
    # (that is, for all samples) to gain a higher-level perspective on how the model makes predictions.
    #
    # There are three primary ways to compute global feature importance values:
    # - Gain: This classic approach introduced by Leo Breiman in 1984 uses the total reduction of loss or impurity
    #   contributed by all splits for a given feature. The motivation is largely heuristic, but it is a commonly used
    #   method to select features.
    # - Split count: This is an alternative approach that counts how often a feature is used to make a split decision,
    #   based on the selection of features for this purpose based on the resultant information gain.
    # - Permutation: This approach randomly permutes the feature values in a test set and measures how much the model's
    #   error changes, assuming that an important feature should create a large increase in the prediction error.
    #   Different permutation choices lead to alternative implementations of this basic approach.
    #
    # All gradient boosting implementations provide feature-importance scores after training as a model attribute.
    # The `lightGBM` library provides five versions, as shown in the following list:
    # - total_gain and gain as its average per split
    # - total_cover as the number of samples per split when a feature was used
    # - weight as the split count from preceding values
    #
    # These values are available using the trained model's .feature_importance() method with the corresponding
    # importance_type parameter. For the best performing XGBoost model, the results are as follows (the total measures
    # have a correlation of 0.8, as do cover and total_cover):
    #
    # While the indicators for different months and years dominate, the most recent 1m return is the second-most
    # important feature from a total_gain perspective, and is used frequently according to the weight measure, but
    # produces low average gains as it is applied to relatively few instances on average).

    def get_feature_importance(model, importance_type="split"):
        fi = pd.Series(
            model.feature_importance(importance_type=importance_type), index=model.feature_name()
        )
        return fi / fi.sum()

    feature_importance = (
        get_feature_importance(lgb_model)
        .to_frame("Split")
        .join(get_feature_importance(lgb_model, "gain").to_frame("Gain"))
    )

    (
        feature_importance.nlargest(20, columns="Gain")
        .sort_values("Gain", ascending=False)
        .plot.bar(subplots=True, layout=(2, 1), figsize=(14, 6), legend=False, sharey=True, rot=0)
    )
    plt.suptitle("Normalized Importance (Top 20 Features)", fontsize=14)
    plt.tight_layout()
    plt.subplots_adjust(top=0.9)
    plt.savefig("images/07-01.png")

    ## Partial Dependence Plots
    # In addition to the summary contribution of individual features to the model's prediction, partial dependence
    # plots visualize the relationship between the target variable and a set of features. The nonlinear nature of
    # gradient boosting trees causes this relationship to depends on the values of all other features.
    # Hence, we will marginalize these features out. By doing so, we can interpret the partial dependence as the
    # expected target response. We can visualize partial dependence only for individual features or feature pairs.
    # The latter results in contour plots that show how combinations of feature values produce different predicted
    # probabilities, as shown in the following code:

    # Requires running cross-validation in [sklearn_gbm_tuning](02_sklearn_gbm_tuning.ipynb).
    class OneStepTimeSeriesSplit:
        pass

    gb_clf = joblib.load("../data/ch12/baseline/sklearn_gbm_model.joblib")

    def get_data(start="2000", end="2018", holding_period=1, dropna=False):
        idx = pd.IndexSlice
        target = f"target_{holding_period}m"
        with pd.HDFStore(DATA_STORE) as store:
            df = store["engineered_features"]

        if start is not None and end is not None:
            df = df.loc[idx[:, start:end], :]
        if dropna:
            df = df.dropna()

        y = (df[target] > 0).astype(int)
        X = df.drop([c for c in df.columns if c.startswith("target")], axis=1)
        return y, X

    def factorize_cats(df, cats=None):
        if cats is None:
            cats = ["sector"]
        cat_cols = ["year", "month", "age", "msize"] + cats
        for cat in cats:
            df[cat] = pd.factorize(df[cat])[0]
        df.loc[:, cat_cols] = df.loc[:, cat_cols].fillna(-1)
        return df

    y_clean, features_clean = get_data(dropna=True)
    X = factorize_cats(features_clean).drop(["year", "month"], axis=1)

    ### 2D Partial Dependency
    # The below plot shows the dependence of the probability of a positive return over the next month given the range
    # of values for lagged 1-month and 3-month returns after eliminating outliers at the [5%, 95%] percentiles.
    # The month_9 variable is a dummy variable, hence the step-function-like plot.

    fig, axes = plt.subplots(nrows=2, ncols=2, figsize=(12, 10))

    X_ = features_clean
    plot_partial_dependence(
        estimator=gb_clf,
        X=X_,
        features=["return_12m", "return_6m", "CMA", ("return_12m", "return_6m")],
        percentiles=(0.05, 0.95),
        n_jobs=-1,
        n_cols=2,
        response_method="decision_function",
        grid_resolution=250,
        ax=axes,
    )

    for i, j in product([0, 1], repeat=2):
        if i != 1 or j != 0:
            axes[i][j].xaxis.set_major_formatter(FuncFormatter(lambda y, _: "{:.0%}".format(y)))

    axes[1][1].yaxis.set_major_formatter(FuncFormatter(lambda y, _: "{:.0%}".format(y)))

    axes[0][0].set_ylabel("Partial Dependence")
    axes[1][0].set_ylabel("Partial Dependence")
    axes[0][0].set_xlabel("12-Months Return")
    axes[0][1].set_xlabel("6-Months Return")
    axes[1][0].set_xlabel("Conservative Minus Aggressive")

    axes[1][1].set_xlabel("12-Month Return")
    axes[1][1].set_ylabel("6-Months Return")
    fig.suptitle("Partial Dependence Plots", fontsize=16)
    fig.tight_layout()
    fig.subplots_adjust(top=0.95)
    plt.savefig("images/07-02.png")

    ### 3D Partial Dependency
    # We can also visualize the dependency in 3D, as shown in the following code that produces the below 3D plot of
    # the partial dependence of the 1-month return direction on lagged 1-month and 3-months returns:

    targets = ["return_12m", "return_6m"]
    pdp, axes = partial_dependence(estimator=gb_clf, features=targets, X=X_, grid_resolution=100)

    XX, YY = np.meshgrid(axes[0], axes[1])
    Z = pdp[0].reshape(list(map(np.size, axes))).T

    fig = plt.figure(figsize=(14, 8))
    ax = Axes3D(fig)
    surface = ax.plot_surface(XX, YY, Z, rstride=1, cstride=1, cmap=plt.cm.BuPu, edgecolor="k")
    ax.set_xlabel("12-Month Return")
    ax.set_ylabel("6-Month Return")
    ax.set_zlabel("Partial Dependence")
    ax.view_init(elev=22, azim=30)
    ax.yaxis.set_major_formatter(FuncFormatter(lambda y, _: "{:.0%}".format(y)))
    ax.xaxis.set_major_formatter(FuncFormatter(lambda y, _: "{:.0%}".format(y)))

    fig.colorbar(surface)
    fig.suptitle("Partial Dependence by 6- and 12-month Returns", fontsize=16)
    fig.tight_layout()
    plt.savefig("images/07-03.png")

    ## SHAP Values
    # At the 2017 NIPS conference, Scott Lundberg and Su-In Lee from the University of Washington presented a new and
    # more accurate approach to explaining the contribution of individual features to the output of tree ensemble
    # models called [SHapley Additive exPlanations](https://github.com/slundberg/shap), or SHAP values.
    # This new algorithm departs from the observation that feature-attribution methods for tree ensembles, such as the
    # ones we looked at earlier, are inconsistent—that is, a change in a model that increases the impact of a feature
    # on the output can lower the importance values for this feature.
    # SHAP values unify ideas from collaborative game theory and local explanations, and have been shown to be
    # theoretically optimal, consistent, and locally accurate based on expectations. Most importantly, Lundberg and Lee
    # have developed an algorithm that manages to reduce the complexity of computing these model-agnostic, additive
    # feature-attribution methods from O(TLDM) to O(TLD2), where T and M are the number of trees and features,
    # respectively, and D and L are the maximum depth and number of leaves across the trees.
    # This important innovation permits the explanation of predictions from previously intractable models with thousands
    # of trees and features in a fraction of a second. An open source implementation became available in late 2017 and
    # is compatible with XGBoost, LightGBM, CatBoost, and sklearn tree models.
    # Shapley values originated in game theory as a technique for assigning a value to each player in a collaborative
    # game that reflects their contribution to the team's success. SHAP values are an adaptation of the game theory
    # concept to tree-based models and are calculated for each feature and each sample. They measure how a feature
    # contributes to the model output for a given observation. For this reason, SHAP values provide differentiated
    # insights into how the impact of a feature varies across samples, which is important given the role of interaction
    # effects in these nonlinear models.

    ### Summary Plot
    # To get a high-level overview of the feature importance across a number of samples, there are two ways to plot
    # the SHAP values: a simple average across all samples that resembles the global feature-importance measures
    # computed previously (as shown in the second plot), or a scatter graph to display the impact of every feature
    # for every sample (as shown in first plot).
    # The scatter plot sorts features by their total SHAP values across all samples, and then shows how each feature
    # impacts the model output as measured by the SHAP value as a function of the feature's value, represented by its
    # color, where red represents high and blue represents low values relative to the feature's range.
    # They are very straightforward to produce using a trained model of a compatible library and matching input data,
    # as shown in the following code:

    X = data[features].sample(n=1000)
    # load JS visualization code to notebook
    shap.initjs()

    # explain the model's predictions using SHAP values
    explainer = shap.TreeExplainer(lgb_model)
    shap_values = explainer.shap_values(X=X)

    shap.summary_plot(shap_values, X, show=False)
    plt.tight_layout()
    plt.savefig("images/07-04.png")

    shap.summary_plot(shap_values, X, plot_type="bar", show=False)
    plt.tight_layout()
    plt.savefig("images/07-05.png")

    ### Feature Interaction
    # ### Force Plots
    # The following force plot shows the cumulative impact of various features and their values on the model output,
    # which in this case was 0.6, quite a bit higher than the base value of 0.13 (the average model output over the
    # provided dataset).
    # Features highlighted in red increase the output. The month being October is the most important feature and
    # increases the output from 0.338 to 0.537, whereas the year being 2017 reduces the output.
    # Hence, we obtain a detailed breakdown of how the model arrived at a specific prediction:

    i = randint(0, len(X))
    # visualize the first prediction's explanation
    shap.force_plot(explainer.expected_value, shap_values[i, :], X.iloc[i, :])

    # We can also compute force plots for numerous data points or predictions at a time and use a clustered
    # visualization to gain insights into how prevalent certain influence patterns are across the dataset.
    # The following plot shows the force plots for the first 1,000 observations rotated by 90 degrees, stacked
    # horizontally, and ordered by the impact of different features on the outcome for the given observation.
    # The implementation uses hierarchical agglomerative clustering of data points on the feature SHAP values to
    # identify these patterns, and displays the result interactively for exploratory analysis, as shown in the following code:

    shap.force_plot(explainer.expected_value, shap_values[:1000, :], X.iloc[:1000])

    ### Interaction Plot
    # Lastly, SHAP values allow us to gain additional insights into the interaction effects between different features
    # by separating these interactions from the main effects. The shap.dependence_plot  can be defined as follows.
    # It displays how different values for 1-month returns (on the x-axis) affect the outcome (SHAP value on the y-axis),
    # differentiated by 3-month returns:
    shap.dependence_plot(
        ind="r01",
        shap_values=shap_values,
        features=X,
        interaction_index="r05",
        title="Interaction between 1- and 5-Day Returns",
    )

    # SHAP values provide granular feature attribution at the level of each individual prediction, and enable much
    # richer inspection of complex models through (interactive) visualization. The SHAP summary scatterplot displayed
    # at the beginning of this section offers much more differentiated insights than a global feature-importance bar
    # chart. Force plots of individual clustered predictions allow for more detailed analysis, while SHAP dependence
    # plots capture interaction effects and, as a result, provide more accurate and detailed results than partial
    # dependence plots.
    # The limitations of SHAP values, as with any current feature-importance measure, concern the attribution of
    # the influence of variables that are highly correlated because their similar impact could be broken down in arbitrary ways.
